Refactor file storage into modular structure
• Split monolithic file.rs into separate modules
• Move SyncMode and utilities to common.rs
• Extract FileConsumerGroup to consumer_group.rs
• Rename file.rs to topic_log.rs for clarity
• Add proper module exports in mod.rs
